feat: dna_abacus + refactor for latency, jitter
